+++ build-trycp.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+# build TryCP server
+cd crates/trycp_server
+cargo build --release --target-dir target