+++ ts/test/trycp/scenario.ts
@@ -1,14 +1,16 @@
 import { AppSignal, AppSignalCb, EntryHash } from "@holochain/client";
 import { URL } from "node:url";
 import test from "tape-promise/tape.js";
-import { stopAllTryCpServers } from "../../src/index.js";
 import { TryCpScenario } from "../../src/trycp/conductor/scenario.js";
 import {
   TRYCP_SERVER_HOST,
   TRYCP_SERVER_PORT,
   TryCpServer,
+  stopAllTryCpServers,
 } from "../../src/trycp/trycp-server.js";
 import { FIXTURE_HAPP_URL } from "../fixture/index.js";
+import { spawnSignalingServer } from "../../src/common.js";
+import { awaitDhtSync } from "../../src/util.js";
 
 const SERVER_URL = new URL(`ws://${TRYCP_SERVER_HOST}:${TRYCP_SERVER_PORT}`);
 
@@ -16,6 +18,8 @@ test("TryCP Scenario - create a conductor", async (t) => {
   const tryCpServer = await TryCpServer.start();
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   const client = await scenario.addClient(SERVER_URL);
   t.ok(client, "client set up");
 
@@ -30,6 +34,8 @@ test("TryCP Scenario - install a hApp to 1 conductor with 1 agent", async (t) =>
   const tryCpServer = await TryCpServer.start();
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   const client = await scenario.addClient(SERVER_URL);
   t.ok(client, "client set up");
 
@@ -56,6 +62,8 @@ test("TryCP Scenario - install a hApp to 2 conductors with 1 agent each", async 
   const tryCpServer2 = await TryCpServer.start(serverPort2);
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   const client1 = await scenario.addClient(serverUrl1);
   const client2 = await scenario.addClient(serverUrl2);
   t.ok(client1, "client 1 set up");
@@ -96,6 +104,8 @@ test("TryCP Scenario - list everything", async (t) => {
   const tryCpServer = await TryCpServer.start();
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   const client = await scenario.addClient(SERVER_URL);
 
   const alice = await scenario.addPlayerWithApp(client, {
@@ -127,6 +137,8 @@ test("TryCP Scenario - receive signals with 2 conductors", async (t) => {
   const tryCpServer = await TryCpServer.start();
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   const client = await scenario.addClient(SERVER_URL);
 
   let signalHandlerAlice: AppSignalCb | undefined;
@@ -184,6 +196,8 @@ test("TryCp Scenario - create and read an entry, 2 conductors", async (t) => {
   const tryCpServer = await TryCpServer.start();
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   const client = await scenario.addClient(SERVER_URL);
 
   const [alice, bob] = await scenario.addPlayersWithApps(client, [
@@ -216,6 +230,8 @@ test("TryCP Scenario - conductor maintains data after shutdown and restart", asy
   const tryCpServer = await TryCpServer.start();
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   const client = await scenario.addClient(SERVER_URL);
 
   const [alice, bob] = await scenario.addPlayersWithApps(client, [
@@ -279,6 +295,8 @@ test("TryCP Scenario - connect to multiple clients by passing a list of URLs", a
   }
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   await scenario.addClientsPlayers(serverUrls);
   t.ok(
     scenario.clients.length === numberOfServers,
@@ -310,6 +328,8 @@ test("TryCP Scenario - create multiple conductors for multiple clients", async (
   }
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   await scenario.addClientsPlayers(serverUrls, {
     numberOfConductorsPerClient,
   });
@@ -348,6 +368,8 @@ test("TryCP Scenario - create multiple agents for multiple conductors for multip
   }
 
   const scenario = new TryCpScenario();
+  [scenario.signalingServerProcess, scenario.signalingServerUrl] =
+    await spawnSignalingServer();
   const clientsPlayers = await scenario.addClientsPlayers(serverUrls, {
     numberOfConductorsPerClient,
     numberOfAgentsPerConductor,
@@ -374,7 +396,7 @@ test("TryCP Scenario - create multiple agents for multiple conductors for multip
     fn_name: "create",
     payload: content,
   });
-  await scenario.awaitDhtSync(alice.cells[0].cell_id);
+  await awaitDhtSync([alice.conductor, bob.conductor], alice.cells[0].cell_id);
   const readContent = await bob.cells[0].callZome<typeof content>({
     zome_name: "coordinator",
     fn_name: "read",